feat: firebase intergration 1
