+++ .gitignore
@@ -164,7 +164,7 @@ key/key.json
 key/vocal-entity-420406-b9648ba69fca.json
 key/llama-433214-97936398ef7a.json
 !key/*.md
-backend/functions/key/vocal-entity-420406-b9648ba69fca.json
+key/WelfareBot.pem
 
 # test audio
 public/*
