--- backend/.firebaserc
@@ -1,5 +1,0 @@
-{
-  "projects": {
-    "default": "welfarebot-7cbc3"
-  }
-}
