--- backend/functions/index.js
@@ -1,4 +1,0 @@
-const functions = require("firebase-functions");
-const app = require("./server");
-
-exports.app = functions.https.onRequest(app);
