+++ backend/stt.js
